backend/pyproject.toml
[tool.poetry]
name = "buy-armenian"
version = "0.1.0"
description = ""
authors = ["Arman <armanghandilyan977@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
django-environ = "^0.11.2"
django-ninja = "^1.1.0"
psycopg = "^3.1.18"
psycopg2-binary = "^2.9.9"
django-countries = "^7.6.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.0"
pytest-django = "^4.8.0"
pytest="^8.2.1"
faker = "^22.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.base"
python_files = "tests.py test_*.py *_tests.py"
filterwarnings = "ignore::django.utils.deprecation.RemovedInDjango51Warning"


